+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.1.0-alpha.8",
-  "contentSha256_12": "76f510da535d",
+  "pkgVersion": "1.1.0-alpha.9",
+  "contentSha256_12": "0bddf097a6be",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.1.0-alpha.8";
-export const contentSha256_12 = "76f510da535d";
+export const pkgVersion = "1.1.0-alpha.9";
+export const contentSha256_12 = "0bddf097a6be";
 export const file = "nodalia-cards.bundle.js";